+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Migrate alert facility change sheet to use Mobius loop
 - Bump AGP to v8.2.0
 - Make entire card scrollable in `OnboardingConsentScreen`
+- Show data protection consent dialog after entering app lock pin for existing users
 
 ### Fixes